+++ include/point_correspondeces.hpp
@@ -7,16 +7,16 @@ using namespace std;
 using namespace cv;
 
 struct Point_Correspondences {
-  explicit Point_Correspondences(const vector<Point2f> &image_points,
-                                 const vector<Point2f> &display_points);
-  vector<Point2f> points_image;
-  vector<Point2f> points_display;
+  explicit Point_Correspondences(const vector<Point2d> &image_points,
+                                 const vector<Point2d> &display_points);
+  vector<Point2d> points_image;
+  vector<Point2d> points_display;
 };
 
 struct Line_Correspondences {
   // Line_Correspondences()
-  vector<vector<Point2f>> pointlines_image;
-  vector<vector<Point2f>> pointlines_display;
+  vector<vector<Point2d>> pointlines_image;
+  vector<vector<Point2d>> pointlines_display;
 };
 
 Line_Correspondences
